170919
12:22PM

I removed the existing table layout for the application in order to see if I can make the viewport cover the entire screen then have the buttons / data be shown in front as overlayed components.

I want to be able to turn on and off layers by creating components for keys. That is the step I'm taking at the moment.