171011
3:06PM

Can now click on an object in the viewport, select it, and show it's user data properties in a table under "Data Table". Problem is that I cannot adjust the look to show the data in a transverse position as well as viewing data objects within data objects. It will only display [object object] in the cell.